% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shcs_dir.R
\name{get_shcs_dir}
\alias{get_shcs_dir}
\alias{set_shcs_dir}
\title{Get/set the default directory of the SHCS data}
\usage{
get_shcs_dir()

set_shcs_dir(x)
}
\arguments{
\item{x}{chr. Path to the SHCS directory.}
}
\description{
Get/set the default directory of the SHCS data
}
